+++ tests/Unit/MethodsTest.php
@@ -4,9 +4,25 @@ namespace SebastiaanLuca\PipeOperator\Tests\Unit\Classes;
 
 use Closure;
 use PHPUnit\Framework\TestCase;
+use SebastiaanLuca\PipeOperator\Item;
 
 class MethodsTest extends TestCase
 {
+    /**
+     * @test
+     */
+    public function it can transform using the static constructor(): void
+    {
+        $result = Item::from('https://blog.github.com')
+            ->parse_url()
+            ->end()
+            ->explode('.', PIPED_VALUE)
+            ->reset()
+            ->get();
+
+        $this->assertSame('blog', $result);
+    }
+
     /**
      * @test
      */